Move sgx-sys to iocuddle-sgx
Rework the crate to export only the ioctl interfaces.
